+++ seatunnel-engine/seatunnel-engine-core/src/main/java/org/apache/seatunnel/engine/core/dag/actions/PartitionTransformAction.java
@@ -28,7 +28,7 @@ public class PartitionTransformAction extends AbstractAction {
 
     private final PartitionSeaTunnelTransform partitionTransformation;
 
-    public PartitionTransformAction(int id,
+    public PartitionTransformAction(long id,
                                     @NonNull String name,
                                     @NonNull List<Action> upstreams,
                                     @NonNull PartitionSeaTunnelTransform partitionTransformation,
@@ -37,7 +37,7 @@ public class PartitionTransformAction extends AbstractAction {
         this.partitionTransformation = partitionTransformation;
     }
 
-    public PartitionTransformAction(int id,
+    public PartitionTransformAction(long id,
                                     @NonNull String name,
                                     @NonNull PartitionSeaTunnelTransform partitionTransformation,
                                     @NonNull List<URL> jarUrls) {
