+++ seatunnel-engine/seatunnel-engine-core/src/main/java/org/apache/seatunnel/engine/core/dag/actions/TransformChainAction.java
@@ -29,7 +29,7 @@ public class TransformChainAction<T> extends AbstractAction {
     private static final long serialVersionUID = -340174711145367535L;
     private final List<SeaTunnelTransform<T>> transforms;
 
-    public TransformChainAction(int id,
+    public TransformChainAction(long id,
                                 @NonNull String name,
                                 @NonNull List<Action> upstreams,
                                 @NonNull List<URL> jarUrls,
@@ -38,7 +38,7 @@ public class TransformChainAction<T> extends AbstractAction {
         this.transforms = transforms;
     }
 
-    public TransformChainAction(int id,
+    public TransformChainAction(long id,
                                 @NonNull String name,
                                 @NonNull List<URL> jarUrls,
                                 @NonNull List<SeaTunnelTransform<T>> transforms) {
