+++ seatunnel-engine/seatunnel-engine-server/src/main/java/org/apache/seatunnel/engine/server/dag/execution/ExecutionVertex.java
@@ -25,7 +25,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class ExecutionVertex {
-    private Integer vertexId;
+    private Long vertexId;
     private Action action;
     private int parallelism;
 }
